+++ tests/AbacTest.php
@@ -16,6 +16,7 @@ class AbacTest extends \PHPUnit_Framework_TestCase
 
     public function testEnforce()
     {
+        $countries = include('tests/fixtures/countries.php');
         $visas = include('tests/fixtures/visas.php');
         $users = include('tests/fixtures/users.php');
         $vehicles = include('tests/fixtures/vehicles.php');