+++ tests/Manager/ConfigurationManagerTest.php
@@ -14,7 +14,7 @@ class ConfigurationManagerTest extends \PHPUnit_Framework_TestCase {
     public function testParseConfigurationFile() {
         $this->manager->parseConfigurationFile([__DIR__.'/../fixtures/policy_rules.yml']);
         
-        $this->assertCount(4, $this->manager->getAttributes());
+        $this->assertCount(5, $this->manager->getAttributes());
         $this->assertCount(4, $this->manager->getRules());
     }
 }